requirements.txt
asyncio~=3.4.3
openai~=1.63.2
python-dotenv~=1.0.1
aiohttp~=3.8.1
